+++ src/hooks/useLogin.ts
@@ -45,7 +45,6 @@ export const useLogin=()=>{
       setIsLoading(true);
       setLoginError(false); // Reset error state before making request
       
-
       try {
         const responseLogin = await login(values.email, values.password);
